+++ client/web/CHANGELOG.md
@@ -76,4 +76,5 @@
 
 - [Divide content into available tabs & finalize](https://inovretail.visualstudio.com/PRODUCT_DEVELOPMENT/_boards/board/t/PRODUCT_DEVELOPMENT%20Team/Stories/?workitem=15811)
 
+- [Physical Evaluation mobile - header](https://dev.azure.com/inovretail/PRODUCT_DEVELOPMENT/_boards/board/t/PRODUCT_DEVELOPMENT%20Team/Stories/?workitem=16001)
 - [Physical Evaluation mobile - form](https://dev.azure.com/inovretail/PRODUCT_DEVELOPMENT/_boards/board/t/PRODUCT_DEVELOPMENT%20Team/Stories/?workitem=16000)